requirements.txt
torch
torchvision
numpy
opencv-python
scipy
Pillow
tqdm
matplotlib